+++ pytos/common/functions/utils.py
@@ -1,4 +1,3 @@
-
 import collections
 import hashlib
 import itertools